+++ common/src/main/java/edu/snu/nemo/common/ir/vertex/executionproperty/ResourceLocalityProperty.java
@@ -18,7 +18,8 @@ package edu.snu.nemo.common.ir.vertex.executionproperty;
 import edu.snu.nemo.common.ir.executionproperty.VertexExecutionProperty;
 
 /**
- * This property decides whether or not to schedule this vertex only on executors where source data reside.
+ * This property decides whether or not to schedule this vertex only on executors where
+ * source (including intermediate) data reside.
  */
 public final class ResourceLocalityProperty extends VertexExecutionProperty<Boolean> {
   private static final ResourceLocalityProperty SOURCE_TRUE = new ResourceLocalityProperty(true);
